requirements.txt
patool~=1.12
python-dotenv~=1.0.0
mutagen~=1.47.0
librosa~=0.10.2.post1